debug: middleware

+++ src/app/controllers/ArtistasController.ts
@@ -26,7 +26,7 @@ class ArtistasController {
         }
     }
     async store(req: Request, res: Response) {
-        middleware(req, res);
+        if (!middleware(req, res)) return res.status(401).json({error: 'Unauthorized'});
         try {
             const artista : Artista = req.body;
             if (!isArtista(artista)) return res.status(400).json({error: 'Object is not of the type: Artista'});
@@ -37,7 +37,7 @@ class ArtistasController {
         }
     }
     async update(req: Request, res: Response) {
-        middleware(req, res);
+        if (!middleware(req, res)) return res.status(401).json({error: 'Unauthorized'});
         try {
             if (!idValid(req.params.id)) return res.status(400).json({error: 'Invalid ID'});
             const artista : Artista = req.body;
@@ -48,7 +48,7 @@ class ArtistasController {
         }
     }
     async delete(req: Request, res: Response) {
-        middleware(req, res);
+        if (!middleware(req, res)) return res.status(401).json({error: 'Unauthorized'});
         try {
             if (!idValid(req.params.id)) return res.status(400).json({error: 'Invalid ID'});
             const result = await ArtistasRepository.delete(new ObjectId(req.params.id));
