+++ src/app/controllers/GenerosMusicaisController.ts
@@ -26,7 +26,7 @@ class GenerosMusicaisController {
         }
     }
     async store(req: Request, res: Response) {
-        middleware(req, res);
+        if (!middleware(req, res)) return res.status(401).json({error: 'Unauthorized'});
         try {
             const genero : GeneroMusical = req.body;
             if (!isGeneroMusical(genero)) return res.status(400).json({error: 'Object is not of the type: Genero Musical'});
@@ -37,7 +37,7 @@ class GenerosMusicaisController {
         }
     }
     async update(req: Request, res: Response) {
-        middleware(req, res);
+        if (!middleware(req, res)) return res.status(401).json({error: 'Unauthorized'});
         try {
             if (!idValid(req.params.id)) return res.status(400).json({error: 'Invalid ID'});
             const genero : GeneroMusical = req.body;
@@ -48,7 +48,7 @@ class GenerosMusicaisController {
         }
     }
     async delete(req: Request, res: Response) {
-        middleware(req, res);
+        if (!middleware(req, res)) return res.status(401).json({error: 'Unauthorized'});
         try {
             if (!idValid(req.params.id)) return res.status(400).json({error: 'Invalid ID'});
             const result = await GenerosMusicaisRepository.delete(new ObjectId(req.params.id));
